Quick bandaid fix for playlist issues.
Star-ratings for playlists stopped working
(at least for me)
in the context of playlists fairly recently.

Essentially it looks like the track uri
is not in the same place it used to be
in the trackListRow elements pending props.

If this wasn't an issue for anyone else
then let me know and I'll just use this
"fixed" version locally.